Correção método DELETE

+++ src/main/java/com/devsuperior/clienteservice/services/ClientService.java
@@ -8,7 +8,6 @@ import com.devsuperior.clienteservice.services.exceptions.ResourceNotFoundExcept
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -58,10 +57,11 @@ public class ClientService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         try{
             repository.deleteById(id);
-        } catch(EmptyResultDataAccessException e){
-            throw new ResourceNotFoundException("Recurso não encontrado");
         } catch(DataIntegrityViolationException e) {
             throw new DatabaseException("Falha de integridade referencial");
         }
